Remove redundant debug logs. Fix nonce setting. Fix ordering in azure docs.
